requirements_ros.txt
rosdep
rosinstall_generator
wstool
rosinstall
pyyaml
empy
defusedxml
netifaces
rospkg
catkin_pkg
boost
